Updated the Search to the results code is in its own file.
